% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimum_pvalue.R
\name{minP}
\alias{minP}
\title{Minimum p-value (minP): a technique for meta analysis}
\usage{
minP(datalist, stat = FALSE, p.adj = FALSE)
}
\arguments{
\item{datalist}{input datalist}

\item{stat}{to show stat, by default stat=FALSE}

\item{p.adj}{p values adjustment, by default p.adj = FALSE}
}
\value{
p_value

statistic if stat=TRUE
}
\description{
The minP method takes the minimum p-vlaue among the K studies as the test statistic.
It follows a beta distribution with degrees of freedom  \deqn{\alpha = 1} and \deqn{\beta = k} under the null hypothesis.
}
\examples{
dl <- sampleData(p=100)
minP(dl, stat = FALSE,p.adj = FALSE)
dl <- sampleData(p=100)
minP(dl, stat = FALSE,p.adj = 'fdr')
}
