% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_matrix.R
\name{p_matrix}
\alias{p_matrix}
\title{Computing a P values Matrix for a List of Dataframes}
\usage{
p_matrix(datalist, stat_test = FALSE)
}
\arguments{
\item{datalist}{input datalist}

\item{stat_test}{statistical tests applied to each biomarker in data frames.Default is false}
}
\value{
A matrix of P-values

A matrix of statistical tests if stat_test = TRUE
}
\description{
Within a dataset, biomarker measurements has to be all continuous or all categorical.
If each biomarker in a dataset has >2 groups, either ANOVA or Kruskal Wallis test will be chosen.

If each biomarker in a dataset has =2 groups, either t-tests or Wilcoxon rank sum test will be chosen.

If each biomarker in a dataset is categorical, either Fisher's exact test or chi-squared tests will be chosen.
}
\examples{
dl <-sampleData()
p_matrix(dl, stat_test=FALSE)
}
