% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleData.R
\name{sampleData}
\alias{sampleData}
\title{Sample data for R meta analysis package}
\format{
A list of 3 data frames, each of which has 101 columns
It's a list of 3 data frames, each of which has 101 columns. For each data frame, first column indicating group membership,
columns 2 to p+1 representing the biomarker measurements for p biomarkers.
 \describe{
 \item{Dataframe1}{Data1 consists of 200 continuous data samples}
 \item{Dataframe2}{Data2 consists of 150 continous data samples}
 \item{Dataframe3}{Data3 consists of 300 categorical data samples}
 }
}
\usage{
sampleData(p = 100)
}
\arguments{
\item{p}{number of columns(by default p=100)}
}
\value{
A list of 3 data frames.
}
\description{
A simple list of dataframes that is for simple testing and running.
}
\examples{
sampleData()
sampleData(p=200)
}
