% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stouffer.R
\name{stouffer}
\alias{stouffer}
\title{Stoufferer's Method: a technique for meta-analysis.}
\usage{
stouffer(datalist, stat = FALSE, p.adj = FALSE)
}
\arguments{
\item{datalist}{input datalist}

\item{stat}{to show stat, by default stat=FALSE}

\item{p.adj}{p values adjustment, by default p.adj = FALSE}
}
\value{
statistic

p_value
}
\description{
Stouffer's method is used to combine the results from several independence tests
bearing upon the same overall hypothesis (H0). The method sums the inverse normal transformed
p-values.
\deqn{Z_{Stouffer} = \sum\limits_{i=1}^{k} Z_{i}/\sqrt{k}}
}
\examples{
dl <- sampleData(p=100)
stouffer(dl, stat = FALSE,p.adj = FALSE)
dl <- sampleData(p=100)
stouffer(dl, stat = FALSE,p.adj = 'fdr')
}
